change LaneID from usize to u32. savings:
scenarios: 458MB -> 431MB
maps: 676MB -> 638MB
prebaked: 85MB -> 80MB

also fix #163

+++ game/src/common/warp.rs
@@ -104,7 +104,7 @@ fn inner_warp(ctx: &mut EventCtx, app: &mut App, line: &str) -> Option<Transitio
                 let r = app.primary.map.maybe_get_r(RoadID(idx))?;
                 ID::Lane(r.children_forwards[0].0)
             }
-            'l' => ID::Lane(LaneID(idx)),
+            'l' => ID::Lane(LaneID(idx as u32)),
             'i' => ID::Intersection(IntersectionID(idx)),
             'b' => ID::Building(BuildingID(idx as u32)),
             'a' => ID::Area(AreaID(idx)),
